+++ backend/src/main/java/com/festago/admin/presentation/AdminController.java
@@ -1,11 +1,5 @@
 package com.festago.admin.presentation;
 
-import com.festago.auth.annotation.Admin;
-import com.festago.auth.application.AdminAuthService;
-import com.festago.auth.dto.AdminLoginRequest;
-import com.festago.auth.dto.AdminSignupRequest;
-import com.festago.auth.dto.AdminSignupResponse;
-import com.festago.auth.dto.RootAdminInitializeRequest;
 import com.festago.common.exception.BadRequestException;
 import com.festago.common.exception.ErrorCode;
 import com.festago.common.exception.InternalServerException;
@@ -31,8 +25,6 @@ import java.time.ZoneId;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.info.BuildProperties;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -52,7 +44,6 @@ public class AdminController {
     private final FestivalService festivalService;
     private final StageService stageService;
     private final TicketService ticketService;
-    private final AdminAuthService adminAuthService;
     private final SchoolService schoolService;
     private final Optional<BuildProperties> properties;
 
@@ -129,21 +120,6 @@ public class AdminController {
             .build();
     }
 
-    @PostMapping("/login")
-    public ResponseEntity<Void> login(@RequestBody @Valid AdminLoginRequest request) {
-        String token = adminAuthService.login(request);
-        return ResponseEntity.ok().header(HttpHeaders.SET_COOKIE, getCookie(token))
-            .build();
-    }
-
-    private String getCookie(String token) {
-        return ResponseCookie.from("token", token)
-            .httpOnly(true)
-            .secure(true)
-            .path("/")
-            .build().toString();
-    }
-
     @GetMapping("/version")
     public ResponseEntity<String> getVersion() {
         return properties.map(it -> ResponseEntity.ok(it.getTime().atZone(ZoneId.of("Asia/Seoul")).toString()))
@@ -164,20 +140,5 @@ public class AdminController {
     @GetMapping("/info")
     public ResponseEntity<Void> getInfo() {
         throw new BadRequestException(ErrorCode.FOR_TEST_ERROR);
-    }
-
-    @PostMapping("/initialize")
-    public ResponseEntity<Void> initializeRootAdmin(@RequestBody @Valid RootAdminInitializeRequest request) {
-        adminAuthService.initializeRootAdmin(request.password());
-        return ResponseEntity.ok()
-            .build();
-    }
-
-    @PostMapping("/signup")
-    public ResponseEntity<AdminSignupResponse> signupAdminAccount(@RequestBody @Valid AdminSignupRequest request,
-                                                                  @Admin Long adminId) {
-        AdminSignupResponse response = adminAuthService.signup(adminId, request);
-        return ResponseEntity.ok()
-            .body(response);
     }
 }
